+++ app/src/main/java/com/spanishdev/tasklistapp/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.spanishdev.tasklistapp.ui.theme
+package com.spanishdev.tasklistapp.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
